Move test_order under the Test::Unit::TestCase
