+++ tool/lib/test/unit.rb
@@ -1262,30 +1262,4 @@ module Test
   end
 end
 
-module MiniTest # :nodoc: all
-  class Unit
-  end
-end
-
-class MiniTest::Unit::TestCase # :nodoc: all
-  test_order = self.test_order
-  class << self
-    attr_writer :test_order
-    undef test_order
-  end
-  def self.test_order
-    defined?(@test_order) ? @test_order : superclass.test_order
-  end
-  self.test_order = test_order
-  undef run_test
-  RUN_TEST_TRACE = "#{__FILE__}:#{__LINE__+3}:in `run_test'".freeze
-  def run_test(name)
-    progname, $0 = $0, "#{$0}: #{self.class}##{name}"
-    self.__send__(name)
-  ensure
-    $@.delete(RUN_TEST_TRACE) if $@
-    $0 = progname
-  end
-end
-
 Test::Unit::Runner.autorun
